drugapp/requirements.txt
Django==5.0.3
GoogleNews==1.6.13
ipython==8.12.3
nltk==3.8.1
numpy==1.26.4
openai==1.14.2
pandas==2.2.1
protobuf==4.25.3
Requests==2.31.0
transformers==4.38.2
google==3.0.0
google-generativeai==0.4.0
gunicorn==21.2.0
newspaper3k==0.2.8
psycopg2
lxml_html_clean
python-dotenv
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
